Add MaskedSparsePCA and improve SparsePCA with builders
- Implement MaskedSparsePCA to support feature selection via boolean mask
- Add builder pattern for better constructor API on both PCA variants
- Optimize column selection for masked operations
- Improve memory efficiency by filtering columns early in processing
