+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "single_algebra"
-version = "0.2.2-alpha.0"
+version = "0.2.3-alpha.0"
 edition = "2021"
 license-file = "LICENSE.md"
 description = "A linear algebra convenience library for the single-rust library. Can be used externally as well."
@@ -44,7 +44,7 @@ num-traits = "0.2.19"
 rayon = "1.10.0"
 simba = { version = "0.9.0", optional = true }
 smartcore = { version = "0.4", features = ["ndarray-bindings"], optional = true }
-single-svdlib = "0.1.0"
+single-svdlib = "0.2.0"
 parking_lot = "0.12.3"
 petgraph = { version = "0.7.1", features = ["rayon"] }
 rand = "0.9.0"
