+++ src/dimred/pca/sparse_masked/mod.rs
@@ -3,12 +3,98 @@ use crate::NumericOps;
 use anyhow::anyhow;
 use nalgebra_sparse::CsrMatrix;
 use ndarray::{s, Array1, Array2};
-use num_traits::Float;
-use single_svdlib::svd_las2;
-use std::fmt::Debug;
-use std::iter::Sum;
+use single_svdlib::{svd_las2, MaskedCSRMatrix};
 
-pub struct SparsePCA<T>
+pub struct MaskedSparsePCABuilder<T>
+where
+    T: single_svdlib::SvdFloat,
+{
+    n_components: usize,
+    alpha: T,
+    tolerance: T,
+    random_seed: Option<u32>,
+    center: bool,
+    verbose: bool,
+    mask: Vec<bool>,
+}
+
+impl<T> Default for MaskedSparsePCABuilder<T>
+where
+    T: single_svdlib::SvdFloat,
+{
+    fn default() -> Self {
+        Self {
+            n_components: 50,
+            alpha: T::from(1.0).unwrap(),
+            tolerance: T::from(1e-6).unwrap(),
+            random_seed: Some(42),
+            center: true,
+            verbose: false,
+            mask: Vec::new(),
+        }
+    }
+}
+
+impl<T> MaskedSparsePCABuilder<T>
+where
+    T: single_svdlib::SvdFloat,
+{
+    pub fn new() -> Self {
+        Self::default()
+    }
+
+    pub fn n_components(mut self, n_components: usize) -> Self {
+        self.n_components = n_components;
+        self
+    }
+
+    pub fn alpha(mut self, alpha: T) -> Self {
+        self.alpha = alpha;
+        self
+    }
+
+    pub fn tolerance(mut self, tolerance: T) -> Self {
+        self.tolerance = tolerance;
+        self
+    }
+
+    pub fn mask(mut self, mask: Vec<bool>) -> Self {
+        self.mask = mask;
+        self
+    }
+
+    pub fn random_seed(mut self, seed: u32) -> Self {
+        self.random_seed = Some(seed);
+        self
+    }
+
+    pub fn center(mut self, center: bool) -> Self {
+        self.center = center;
+        self
+    }
+
+    pub fn verbose(mut self, verbose: bool) -> Self {
+        self.verbose = verbose;
+        self
+    }
+
+    pub fn build(self) -> MaskedSparsePCA<T> {
+        MaskedSparsePCA {
+            n_components: self.n_components,
+            alpha: self.alpha,
+            tolerance: self.tolerance,
+            random_seed: self.random_seed.unwrap_or(42),
+            components_: None,
+            explained_variance_: None,
+            mean_: None,
+            mask: self.mask,
+            center: self.center,
+            verbose: self.verbose,
+        }
+    }
+}
+
+pub struct MaskedSparsePCA<T>
 where
     T: single_svdlib::SvdFloat,
 {
@@ -19,11 +105,12 @@ where
     components_: Option<Array2<T>>,
     explained_variance_: Option<Array1<T>>,
     mean_: Option<Array1<T>>,
+    mask: Vec<bool>,
     center: bool,
     verbose: bool,
 }
 
-impl<T> SparsePCA<T>
+impl<T> MaskedSparsePCA<T>
 where
     T: single_svdlib::SvdFloat + NumericOps + 'static,
 {
@@ -32,6 +119,7 @@ where
         alpha: T,
         tollerance: Option<T>,
         random_seed: Option<u32>,
+        mask: Vec<bool>,
         center: bool,
         verbose: bool,
     ) -> Self {
@@ -43,6 +131,7 @@ where
             components_: None,
             explained_variance_: None,
             mean_: None,
+            mask,
             center,
             verbose,
         }
@@ -50,28 +139,31 @@ where
 
     pub fn initialize_components(&self, x: &CsrMatrix<T>) -> anyhow::Result<Array2<T>> {
         let n_samples = x.nrows();
-        let n_features = x.ncols();
-
-        if self.n_components >= n_features {
+        if x.ncols() != self.mask.len() {
             return Err(anyhow!(
-                "Number of components ({}) must be less than number of features ({})!",
-                self.n_components,
-                n_features
+                "The mask vector length and the number of features (columns) have to be the same!"
             ));
+        }
+        let mut n_features = 0usize;
+        for f in self.mask.clone() {
+            if f {
+                n_features += 1;
+            }
         }
 
         let optimal_iterations = n_samples.max(n_features);
-        let svd = svd_las2(
-            x,
+        let masked_matrix = MaskedCSRMatrix::new(x, self.mask.clone());
+        let svd_masked = svd_las2(
+            &masked_matrix,
             self.n_components,
             optimal_iterations,
             &[T::from(-1.0e-30).unwrap(), T::from(1.0e30).unwrap()],
             T::from(10e-6).unwrap(),
             self.random_seed,
         )
-        .map_err(|e| anyhow!("SVD computation failed: {}", e))?;
+        .map_err(|e| anyhow!("SVD computation failed {}", e))?;
 
-        let components = svd.vt.slice(s![..self.n_components, ..]).to_owned();
+        let components = svd_masked.vt.slice(s![..self.n_components, ..]).to_owned();
 
         if self.verbose {
             println!("Dimensionality reduction summary:");
@@ -91,7 +183,19 @@ where
 
     pub fn fit(&mut self, x: &CsrMatrix<T>, max_iter: Option<usize>) -> anyhow::Result<&mut Self> {
         let n_samples = x.nrows();
-        let n_features = x.ncols();
+        if x.ncols() != self.mask.len() {
+            return Err(anyhow!(
+                "The mask vector length and the number of features (columns) have to be the same!"
+            ));
+        }
+        let mut n_features = 0usize;
+        let mut cols_to_use: Vec<usize> = Vec::new();
+        for (ind, &val) in self.mask.iter().enumerate() {
+            if val {
+                n_features += 1;
+                cols_to_use.push(ind);
+            }
+        }
         let n_t_samples = T::from(n_samples).unwrap();
 
         if self.center {
@@ -108,7 +212,6 @@ where
         }
 
         let mut components = self.initialize_components(x)?;
-
         let max_iter = max_iter.unwrap_or(1000);
 
         let mut converged = false;
@@ -125,30 +228,32 @@ where
             for (row_idx, row) in x.row_iter().enumerate() {
                 for k in 0..self.n_components {
                     let mut score = T::zero();
-                    for &col_idx in x.col_indices() {
+                    for (idx, &col_idx) in cols_to_use.iter().enumerate() {
                         let val = row.get_entry(col_idx).unwrap().into_value();
                         let effective_val = if self.center {
                             val - mean[col_idx]
                         } else {
                             val
                         };
-                        score += effective_val * components[[k, col_idx]];
+                        score += effective_val * components[[k, idx]];
                     }
                     u[[row_idx, k]] = score;
                 }
             }
 
             for j in 0..n_features {
+                let col_idx = cols_to_use[j];
                 for k in 0..self.n_components {
                     let mut loading = T::zero();
 
                     for (row_idx, row) in x.row_iter().enumerate() {
-                        if let Some(val) = row.get_entry(j) {
+                        if let Some(val) = row.get_entry(col_idx) {
                             let effective_val = if self.center {
-                                val.into_value() - mean[j]
+                                val.into_value() - mean[col_idx]
                             } else {
                                 val.into_value()
                             };
+
                             loading += effective_val * u[[row_idx, k]];
                         }
                     }
@@ -199,6 +304,20 @@ where
     }
 
     pub fn transform(&self, x: &CsrMatrix<T>) -> anyhow::Result<Array2<T>> {
+        let n_samples = x.nrows();
+        if x.ncols() != self.mask.len() {
+            return Err(anyhow!(
+                "The mask vector length and the number of features (columns) have to be the same!"
+            ));
+        }
+
+        let mut cols_to_use: Vec<usize> = Vec::new();
+        for (ind, &val) in self.mask.iter().enumerate() {
+            if val {
+                cols_to_use.push(ind);
+            }
+        }
+
         let components = self
             .components_
             .as_ref()
@@ -208,25 +327,25 @@ where
             .as_ref()
             .ok_or_else(|| anyhow!("Must be fitted before transform!"))?;
 
-        let n_samples = x.nrows();
         let mut transformed = Array2::zeros((n_samples, self.n_components));
 
         for (row_idx, row) in x.row_iter().enumerate() {
             for k in 0..self.n_components {
                 let mut score = T::zero();
-                for &col_idx in x.col_indices() {
-                    let val = row.get_entry(col_idx).unwrap().into_value();
+                for (cidx, &col_idx) in cols_to_use.iter().enumerate() {
+                    let val = row
+                        .get_entry(col_idx)
+                        .map_or(T::zero(), |entry| entry.into_value());
                     let effective_val = if self.center {
                         val - mean[col_idx]
                     } else {
                         val
                     };
-                    score += effective_val * components[[k, col_idx]];
+                    score += effective_val * components[[k, cidx]];
                 }
                 transformed[[row_idx, k]] = score;
             }
         }
-
         Ok(transformed)
     }
 
@@ -268,86 +387,5 @@ where
     ) -> anyhow::Result<Array2<T>> {
         self.fit(x, max_iter)?;
         self.transform(x)
-    }
-}
-
-pub struct SparsePCABuilder<T>
-where
-    T: single_svdlib::SvdFloat,
-{
-    n_components: usize,
-    alpha: T,
-    tolerance: T,
-    random_seed: Option<u32>,
-    center: bool,
-    verbose: bool,
-}
-
-impl<T> Default for SparsePCABuilder<T>
-where
-    T: single_svdlib::SvdFloat,
-{
-    fn default() -> Self {
-        Self {
-            n_components: 50,
-            alpha: T::from(1.0).unwrap(),
-            tolerance: T::from(1e-6).unwrap(),
-            random_seed: Some(42),
-            center: true,
-            verbose: false,
-        }
-    }
-}
-
-impl<T> SparsePCABuilder<T>
-where
-    T: single_svdlib::SvdFloat,
-{
-    pub fn new() -> Self {
-        Self::default()
-    }
-
-    pub fn n_components(mut self, n_components: usize) -> Self {
-        self.n_components = n_components;
-        self
-    }
-
-    pub fn alpha(mut self, alpha: T) -> Self {
-        self.alpha = alpha;
-        self
-    }
-
-    pub fn tolerance(mut self, tolerance: T) -> Self {
-        self.tolerance = tolerance;
-        self
-    }
-
-    pub fn random_seed(mut self, seed: u32) -> Self {
-        self.random_seed = Some(seed);
-        self
-    }
-
-    pub fn center(mut self, center: bool) -> Self {
-        self.center = center;
-        self
-    }
-
-    pub fn verbose(mut self, verbose: bool) -> Self {
-        self.verbose = verbose;
-        self
-    }
-
-    pub fn build(self) -> SparsePCA<T> {
-        SparsePCA {
-            n_components: self.n_components,
-            alpha: self.alpha,
-            tolerance: self.tolerance,
-            random_seed: self.random_seed.unwrap_or(42),
-            components_: None,
-            explained_variance_: None,
-            mean_: None,
-            center: self.center,
-            verbose: self.verbose,
-        }
     }
 }
